Added methods for phases, rounds and card back-end
doTurn will play all the cards of every player and it will
also activate the board elements. I also move cardMove to
board. In addition i added an owner field to Card so that
we can know who played a given card.
I also added a lot of documentation.
All we're missing is selection of cards for that part to be
complete.
We must finish the conveyors and push methods.